rework Deployment's stop handling
The current implementation of Deployment's stop command was a
hornet's nest. Really. No kidding. It would basically have a 50-50
chance to do something wrong if called outside the normal flow (i.e.
outside of Roby's garbage collection mechanism)

1. STOPPED/RUNNING race condition

The cleanup loop was testing for rtt_state to check if the task
is STOPPED before attempting to clean it up. If, at the same time,
Syskit is itself trying to configure or start it, this races.
It would sometime abort the deployment's stop.

2. Blocking behavior because of task cleanup

The current flow would cleanup the tasks *and* let the orocos.rb's
implementation of Process also clean them up. It's OK-ish in normal
circumstances, but is rather obviously sub-optimal in basically
every non-nominal situation.

Syskit does assume that tasks should already be stopped before
reaching Deployment#stop. This is why we only check for cleaning
up tasks that are in STOPPED state. orocos.rb does not and also
attempts to stop them. This basically means that Process#kill
might block (and block Syskit) for an indefinite amount of time.

However, we do need to cleanup components before we stop the process
in the nominal situation. So, to account for both, the following
has been implemented:

- the stop! command "guesses" whether it is in a nominal and non-nominal
  situation by checking whether the deployment still supports tasks
  in the plan (i.e. whether we're in something that looks like an
  orderly shutdown or an emergency one)
- If there are tasks in the plan, we assume stop! was called to kill
  the deployment quickly (emergency !) and thus we bypass the cleanup.
  We also call kill with the new 'hard' argument to true (which
  internally causes the process to be SIGKILL'ed)
- If it does not, we assume stop! was called in the normal workflow
  and attempt to cleanup (in Syskit, not in orocos.rb)
- A new 'kill' event may be used to force the usage of the kill workflow.