fix TLS tests in test_spawn_server
The first thing was to separate the private keychain (for the
server) from the public one (for the client). The certificates
were created with

  openssl req -x509 -sha256 -nodes -days 3650 -newkey rsa:2048
              -keyout cert-private.crt -out cert.crt
  cat cert.crt >> cert-private.crt

Once this was done, we made sure the certificate's CN matches
the hostname we use to connect, or the validation fails. Since
all of it is in-process and self-signed, we definitely should
be able to avoid this from a security standpoint. However, I
couldn't find a way to disable the hostname validation within
Net::FTP