remote_processes: separate the request for an upload to the actual upload
The current implementation, even though it was spawning a thread for
transfer, was not replying to the client until the transfer was
completed. This was in effect blocking the client.

This commit changes the implementation to separate the request, which
returns immediately, from the completion state, which is now a separate
request.