test: implement together_with argument for assert_can_instanciate and _deploy
ProfileAssertion iterates on the first argument to the assertion, i.e.
tests them in isolation. There was so far no way to test "check that
this or these actions can be deployed/instanciated along with this/these
other actions and compositions.

This can be used e.g. to test that "permanent services" that always
run do not interfere with the deployment of definitions from other
profiles.