include code fo spawning Log Transfer Server in plugin.rb
Creating an object TmpRootCA, it spawns an automatic new
root CA, a public certificate (for the Process Servers)
and private certificate (for the  Spawn Server).
All created within the TmpRootCA class.

It is also defined a user and password from the  randomly
generated password method in the TmpRootCA class.

After that, a method to spawn a FTP Log Transfer Server is
called, provided with the app log directory, user, password
and the private certificate

[tmp_root_ca] Cleans up testing code and adjusts new methods
and attributes

[test_plugin] Includes initial code for unit testing Log Transfer
Spawn Server funcionality