net_gen: when instanciating, set fullfilled_model only with meaningful arguments
"meaningful arguments" are the arguments that are actually declared
in the task model (with the `argument`) property. These arguments
are constrain what can e.g. replace the task, or if the task may
or may not start (when they are unset).

However, it is possible to pass other arguments, which are useful
to the task, but e.g. should not constrain the replacements

However, during instanciation, syskit would set {#fullfilled_model}
explicitely with all arguments given to the InstanceRequirement object
Since an explicit #fullfilled_model takes precedence, these not-meaningful
arguments were affecting replacement checks.