fix cancellation of asynchronous plan generation
The current implementation had a race condition in case the
generation was already in progress. It was queueing the transaction
discard using a execution_engine.once block, but that was not
guaranteed to be executed before the next apply_require_modification
call. When apply_require_modification was called first, it would
apply the changes.

Moreover, the applied transaction would have been generated while
(1) the syskit_current_resolution_keepalive transaction had been
discarded, and (2) while another resolution could have been in
progress.

A.k.a. recipe for disaster.

Simplify the execution flow by only "cancelling" in #cancel, letting
the generation finish (or, in case the future was pending, being
cancelled - which is also transitioning to finished), and handle
the cancellation in Async#apply. This makes sure we follow the same
execution flow in all cases, and guarantees that we serialize the
generation(s).