corrects TmpRootCA Certificates creation dir and other corrections
[log_transfer_integration.rb] It includes a modification in the
TmpRootCA class, introducing as argument a target dir where to create the
temporary Certificates. This implementation was created having 'app.log_dir'
in mind to store the Certificates while the Application is running.
In the clean_up method was added a 'clean_tmp_cert_dir' method for the
removal of this directory and its content when the Application is finished.

[plugin.rb] changes in the TmpRootCA creation command and includes the
'clean_tmp_cert_dir' method in clean_up

[test_plugin.rb] corrects method positioning and removes unused code