refactor reconfiguration tests to avoid over-mocking
The tests where stopping at the needs_reconfiguration? call,
which was tested elsewhere. However, I found a bug that was
related to the integration (order in which things get called),
which can't possibly be caught this way.

Modify the test to avoid mocking altogether, which I think is
a better method in this case.