introduce #update_properties in a backward-compatible way
The current configuration flow had one day be changed to

1. create instance, initialize its properties with the default
   values (read at process startup from the component)
2. read remote properties
3. decide whether to reconfigure: compare properties with
   expected property values
4. call #cleanup
5. call #configure (Ruby), applies configuration on the default values
   read in (1)
6. call #configure (C++)

The obvious mistake here is that in (2), the component's actual
configuration has not been applied. In essence, we are reconfiguring
all components.

There is no way to move the call to the Ruby #configure from before
cleanup. Any component that uses non-idempotent operations to configure
itself would start failing.

This commit introduces `#update_properties` for this purpose: an
idempotent "only update the properties" which is called in all cases
before the reconfiguration check. #configure gets only called if
the component was reconfigured.

To ease migration, the new scheme is used if:
- `update_properties` is explicitly defined on a component model
  (even if it is only calling super)
- `configure` is not defined on a component model

The new behavior can also be enabled globally with

Roby.app.syskit_task_context_uses_update_properties = true