fix dynamic services comparison in Deployment#configuration_changed?
Because of task merges, individual task's models might be different
(because they are specialized models). In this case, the dynamic
services' #== method returns false. We should be using #same_service?
which is meant to ignore this difference, instead looking whether the
services being compared are instances of the same dynamic service model
with the same options.