allow tuning the buffer size for have_(no_)new_sample(s)
This solves issues related to .matching {}, which causes the number
of samples considered by the predicates to be smaller than the actual
number of samples received. Buffers are therefore now needed on the
have_no_new_sample predicate. Moreover, their size can't be predicted
for any of the three existing predicates.

Set a default size of 100, and document the reason for this.