test: add a separate 'live' test run in test/live/
All files there will be executed using `syskit orogen-test`, allowing
them to run actual components. For speed, the rest of the test suite
avoids spawning the process server and loading external orogen models.
This is the best way to write the tests that really require a specially
crafted C++ component.