feat(run): optionally automatically restart deployments with fatal or fail-to-stop tasks
Until now, deployments with components in FATAL_ERROR (or, worse, that
failed to stop) would end up in the plan with no automated way to handle
them. This would lead to essentially needing a human act, and sometimes
would affect the overall system behavior in an unpredictable way.

Previous commits at least made sure we would not be trying to reuse
such components. I.e. we would fail at netgen time and the error would
be loud.

This commit adds an optional mechanism to auto-restart the affected
deployments, if the opportunistic method did not do it already.