fix(run): handle the asynchronicity between exception and fatal
There fatally is a delay between the exception signals and the
fatal error ones. Often, it's little, but some times it is enough
for Syskit to process one before the other. In this case, since
exception is forwarded to stop, the fatal_error event gets lost.

This commit fixes the issue by delaying the exception event until
we receive the task's stop event (and are therefore sure of the
task's actual state)