fix(run): properly handle state reader(s) being disconnected
If this happens, we don't *know* what the state of the task is, we
can't stop or abort it. Put it in quarantine (now that this does
something meaningful) and let the quarantine recovery code deal with
it.

Try, still, to stop the task if it happens while running. This may fail,
but if it works it will make the recovery easier (and the system's
general life simpler)