fix(runtime): fix deadlock in RemoteStateGetter
If a `pause` was called in between `disconnect` and `join`, the `join`
could block forever. This was due to `pause` resetting the run_event
Event, which caused the loop to never check exit_condition.

Set and check both Event under a mutex to avoid the race, and make
sure that #pause raises if the poll thread is being quit. In
TaskContext's stop event handler, ignore a remote state getter
that is being or has been disconnected.