fix(test): spurious test failure in live test
The bug was a conjunction of three things:
- The tests related to losing both the state reader and the state getter
  would have `join_all_waiting_work(true)`. This caused the tests
  to wait for the blocked method (usually stop) to return.
- update_task_states would call TaskContext#validate_state_reader_connected,
  but still read new states.
- The tests would only mock the remote state getter connected? method,
  which caused update_task_states

So, essentially, we had a race condition between the expect_execution
leaving (that is, the blocking method returning) and the deployment_kill
call. In essence, the tests were not testing what we wanted, that is that
the runtime code would stop reading and reacting to state.

The test got changed to join_all_waiting_work(false) until quarantine,
then sleep to let the blocking call return and then check that the
relevant event was not emitted. These tests failed under the current
implementation.

To fix the tests, I changed update_task_states to not process tasks
whose state reading is completely lost.