fix(runtime): ignore finished children in Composition#executable?
Composition#executable? is implemented so that compositions are started
only when all their children are started.

However, if we want to use a child to actually implement an asynchronous
start, this fails: the child emits success, becomes non-executable which
blocks the emission of start.

Modify Composition#executable? to ignore non-pending children, which
are either executable (if running) or will never be (if finished)