fix(droby): allow loading old log files with incompatible types
Newly created log files can be loaded regardless of whether the
orogen model or its depentent types have changed. This was enabled
by a change in log file format.

While the old log files can still be loaded, they still have the
issue that triggered the change in the first place: the inability
to load them if a type or orogen model has changed since.

This commit re-establishes this ability by allowing to use empty
stub models for orogen tasks, which does not force-load types that
are present in the current workspace.