fix: use the localhost process server in --single mode
The current code was trying to connect to a process server
on localhost, but at the process server's configured port. This
really made no sense.

--single is meant to provide us with a way to run a distributed
system on a single machine. What makes the most sense w.r.t.
process servers is to simply use the local one in this case.