feat: add 'read_only' attribute for TaskContext
read-only tasks are tasks that share a resource (i.e "#orocos_task") with another task running on a different Syskit instance.
In that way, it does not interfere with the resource, as it will only perform read operations, skip setup,
and also not stop when interrupted (so child tasks are not garbage collected).