fix(net_gen): remove dead code
The seen_loggers test was always false (and the branch therefore
always taken). I guess it was introduced in a earlier version of
the code, to avoid un-doing new connections created by the code
itself. Since the code now only handle a single deployment (and
not the whole logger), it can be removed - as the new unit tests
assess.

For information, the new unit tests were not passing if I was
updating the seen_loggers