fix(doc): CSS-based scaling of the dataflow and hierarchy graphs
The width/height is (well known ...) the size of the enclosing box.
Moreover, YARD's own CSS sets max-width to 100%.

Use transform to scale, and override max-width. This causes some graphs
to overflow a bit, but we'll deal with that another time (for now,
people can open the files with a right click and "open image")