feat: add wait running command to remote process client and server
This command is used to wait for the process to be running, which means
that the process has received the IOR from the orogen tasks.  The IOR is
then mapped to each task, and used to create the TaskContexts.

The server returns a Hash reporting any exceptions the occured, and
if there wasnt one, it reports the iors for the process' tasks or nil if
no iors were received yet.