fix: make implicit/explicit FTPs configurable
Implicit FTPs support is broken on Ruby 2.5. The current implementation
was switching based on whether the current ruby is 2.5 or 2.7, but that
did not account for mixed systems, where the main syskit would run
one version and process servers another.

Essentially, one should use implicit ftps if all rubies are 2.7+,
and not use it otherwise.