fix: read only tasks not getting the message that the component is already in running state
Simply refactoring the existing tests was enough, but it was mostly
triggered when restarting a read-only task. The state reader would be
"empty", that is the transition to RUNNING was already read, and therefore
the state update poll would never trigger the start event.

We now initialize a task's orogen state with the last known state of
the component if available.