fix: forcefully export types to Types
The flag was present because of the cost of the old constant-based
implementation. This has been fixed a few years back, and exporting
now is virtually free.

Export is still disabled if no type module is provided to the loader
(the default)

Configuration#export_types(?|=) is kept for backward compatibility