fix: heisenbug in remote process server's #reap_dead_subprocesses
We sometimes we get a PID that has no corresponding registered process
and handlee_dead_subprocess returns nil. Do not notify the client about
these processes as it possibly can't do anything about them.

This was causing an exception in announce_dead_processes as it could
not handle the nil.

The reason for this process ID not being known is unknown right now.