fix: handle a delay between the fatal error and the component returning from stop
Syskit has two communication channels with components. One is the return
of state transition commands (start/stop essentially) and one is the
state port.

When a component goes to FATAL while it's being stopped, the fatal might
have been processed before stop gets to return. If a response to the
transition to fatal is to kill the deployment - something we have to
do on our system - it might be that the deployment is down and the
component aborted when stop returns. This is (rather obviously) not
a normal condition in quarantined!, and the method was raising.

Handle this case, and test for it.