fix: allow passing a Syskit model as a deployment task's model
We were currently passing the orogen model, which was resolved
to a Syskit model using TaskContext.model_for

This is based on the obvious assumption that only one Syskit model
may ever have a given orogen model, which is broken by syskit-log.
In addition, it would not protect the syskit models against Ruby's
garbage collector, which is I believe why we were having random
failures in tests for Syskit models that did not exist.