fix: preload all typekits we might need prior to execution
Typekit loading is currently at least partially done at runtime
"at first access", which in the case of syskit is done in multithreaded
contexts. We have had some heisenbugs in CI that were probably caused
by the typekit not yet being available in some codepaths.

Preload all typekits we need in prepare - which actually is quite a
heavy operation, and thus is better done once and for all.