fix: disable listening to configuration changes during tests
We rather obviously don't need it, and it seems to be the source
of many open file descriptors (and thus failures because there
aren't any available anymore).

I say "seems" because the problem does disappear after this change,
but I couldn't figure out why it happened in the first place. The
only things I could figure out are that the file descriptors are
allocated on the C side (they have no associated IO object in Ruby),
are pointing to the bundle's config/orogen/**/* folders and are
very transient (come and go). I did measure a thousand or so of them...
So definitely a possible source of hitting the limit of open file
descriptors.