chore: copy plan marks when applying merge group
This became a necessity as we pretend to do early deployments at network
generation. In this scenario, when a "free" task context is merged into its
deployed counter part its plan marks is not carried out to the deployment task
instance. Then, this task instance would be garbage collected because it did
not carried the permanent mark from the replace "free" task context