fix: working Async along with the task inspector
This commits adapts/adds new calls to please Vizkit. The most important change
is that the initialization of the state values is now done by the port manager
instead of synchronously. This is made necessary as the task inspector registers
a callback on the state change before it does on reachability, leading sometimes
to having the state remaining INITIALIZING