fix: remove requirement for proxy task to start after planner
With the new resolution success event scheme, the planner no longer
finishes when the task was planned. We removed this requirement taking
into consideration that the root task does not permission to run before
the planner was done anyways. This way, we can use the resolution
success event to signal that the planning task resolution was done, but
keep the planning task running. 

PS: we need the planning task to be running so we can safely emit failed
on it without risking failing a non-running task.