feat: signal that the resolution of a instance requirement task is done via a resolution success event
Previously, when an instance requirement task resolution was done, a
success event was emitted. This commit adds a resolution success event
to do that, which isnt a terminal state. A running planning task is now
an indicator that that task should be considered during the work plan
requirement tasks discovery. We also use that to emit failed events only
for running tasks.