chore: split may_merge_task_context?
It was being used for Composition and Placeholder (i.e. non TaskContext) merge
evaluation. This wasn't a problem until know because it tests for execution_agent
existence which is nil for Composition and Placeholder, but know we test for
TaskContext#orocos_name and orocos_name is not a member of Composition neither
Placeholder