fix: race condition in parallel disconnection
Ruby's Hash is *not* thread safe when #hash has been redefined. When the Hash
code calls the hash method, the Ruby scheduler may schedule another thread that
might create some memory corruption if said thread also modifies the same hash. 

Turns out that Concurrent::Map does not work around this.

This is the case in the disconnection codepath, since Orocos::TaskContext#hash
has been modified so that two objects that point to the same remote task look
the same as hash keys. So far, this bug only has had an effect in CI, showing up
as memory corruption and either a segmentation fault or what looks like type
confusion ("an integer does not have method remove_disconnection_from`)

The commit modifies the connection codepath, even though it was not affected (it
uses syskit tasks as keys and not orocos tasks, and the syskit tasks do not have
their #hash method redefined). This is for symmetry - and to make sure the same
problem does not appear along some changes to these codepaths.