fix: race condition in the exception state confirmation code
The orogen 'state reporting' uses ports, and sends state changes before they
happen (i.e. the call to exception() sends the exception state on the port when
it is called, before the various hooks are called).

This is a problem in Roby/Syskit as the exception events are terminal. We
therefore have to wait to be sure that the task has finished its transition
before we process the events. This code is mostly in
TaskContext#update_orogen_state_in_exception.

This commit fixes a race between the final state reported by reading
TaskContext#rtt_state (a remote call) and the state reporting coming from the
port. In some conditions, the former will come before the latter. This was
making some tests fail (and, in general made the state reporting buggy) in cases
where we transition to fatal as we would emit `exception_event` (which came
before from the port) but not `fatal_error_event` (which came only from the
rtt_state call).

This commit makes sure we wait for the final state to be received on the port
before we process everything, and adds a test case for it.