fix: behavior of quarantine recovery strategies
The tests were not "clean" as to which flag
(opportunistic_recovery_from_quarantine and
auto_restart_deployments_with_quarantines) where enabled or not. This led to
tests that should be related to one (auto_restart) actually being dependent on
the other (opportunistic recovery).

The test layout cleanup actually also cleaned up the setting of these flags,
which triggered test failures. This commit resolves them. The main change is the
addition of a way for a transaction to require a deployment to be killed
(schedule_for_kill!), which is used by auto_restart when a deployment that has
quarantines running needs to be stopped.