fix: error reporting on property update errors
Failures during property updates would generate errors in a rather unexpected
way w.r.t. configuration and runtime errors in Syskit. This commit modifies the
handling to make sure that
- failure during setup leads to a fail-to-start condition
- failure during runtime leads to the emission of internal_error_event and the
  task stop

This is done by modifying the error type to be a subclass of the new
Roby::InternalTaskError type, and making sure it is propagated using Roby's
normal error handling mechanisms.