fix: heisenbug in autorestart test
The issue is that the autorestart mechanism now relies on a flag passed from the
transaction to the plan, and a response of the runtime code for this flag.

`syskit_deploy` calls `execute` in some places, so the whole mechanism would
trigger and the following `expect_execution` might not see the emission of the
abort event.

Artificially set the flag to false with a mock during `syskit_deploy` to ensure
the test passes.