fix: race condition while waiting for log completion in tests
The remote process manager's log upload state 'pending' count would be very hard
to get race-free. In this case, the pending count it taken (quite) after the
result queue has been emptied, in a code that looks like

    result = empty_queue
    return SomeObject.new(pending.value, result)

If a result is queued between the two lines (which in practice are much more
than two lines), pending will be decremented but result does not contain the
result.