fix: do a garbage collection run after processing resolution errors
There were tasks in the plan that would be useless after a
resolution error one of the tasks that depended on it (i.e. a service that
is not provided). This removes them.

Also, I adapted the exception that would use the merge solver, since
these tasks are now removed the plan and would therefore we couldnt use
the merge solver during its pretty print (it needs the task's plan)