fix: remove the notion of 'complete' dataset transfer
The whole flag is here to simplify how we separate archives containing pocolog files
from the ones that contain the rest (text and roby log files), with the
objective that we can easily get rid of parts of the pocolog files while making
sure we keep the rest.

This has no meaning whatsoever in data transfer. Plus, the transfer loop was not
re-processing the dataset right away, which added a 10 min delay to get a full
dataset transfer.