fix: do not re-resolve readers that are using plan matchers
We introduced a mechanism to avoid re-resolving queries for readers that have a
working reader already. This mechanism was actually not being triggered for the
very common case of matching a dataservice. This was because a data service's
actual port (the one that is used to create e.g. a reader) is not equal to the
one that is being matched. The matcher's current_selection_valid? method would
then not recognize its argument as valid.

Fix this by returning the "raw" port and letting DynamicPortBinding#update
resolving the actual port. This is more tricky than it looks because we now have
to actively handle the finalize case - which was implicitly handled before since
a plan query cannot return a finalized object.