chore: add tests for dynamic port binding caching behaviour with finalized tasks
Now that we keep the last match and check it with Matcher#===, I wondered
whether finalized tasks have become a special case, which was implicitly handled
before by the call to find_tasks. They are actually not, PlanObjectMatcher#===
returns false in this case.

Just add unit tests to make that support explicit.