perf: use 'orocos_name' to filter candidate lists when merging task contexts
My take on Syskit's current slowness on our systems is that we have a LOT of
GPIO-driving tasks. The merge solver was designed on the basis that filtering
on type would be enough to avoid hitting large task groups. This is not the
case anymore here.

It is doing a N^2 step on more and more tasks, and that takes forever.

When early_deploy is set, however, orocos_name will be set *and* we can only
merge two tasks if they have the same deployment (the same orocos_name). Use
this to filter out merge candidates.

It reduced the complete system generation + apply loop from ~1.5s to 0.5s.