fix: corner-case in template generation due to delayed arguments
Delayed arguments in general, and DelayedArgumentFromState in particular, may
now be transformed into a static value during merge. This is great in the
general state of things, but breaks templates.

This code disables the merge step for the delayed arguments, *except* in the
case of default values, since default values do not have this behaviour *and*
they are extremely common