fix: support multiple exceptions on ProfileAssertionFailed
When the capture_errors_during_network_resolution flag is on, the
ProfileAssertionFailed will receive an array of exceptions (even if its
a single exception). At first glance it wasnt an issue, but
Roby.display_exception expects that the array is flattened, so it would
raise whenever trying to display exceptions with a
ProfileAssertionFailed.