feat: use model to override the output reader instead of the instance
The model is more informative than the instance, and in hindsight is the
better choice to be overriden. For example, by overriding the model we
can change the type of the reader. Keep in mind whoever overrides the
model probably still wants to implement a new instanciated version of it. 
Do so by setting the klass.