fix: use orocos_name only in MergeSolver#may_merge_task_contexts?
The lazy deployments are only settings orocos_name. The objective of the
execution_agent test is to actually check that the task's deployment has been
chosen and is the same (even if the instance is different), which can be done by
comparing orocos_name

Modify it. This fixes merging of tasks with lazy-deploy enabled