fix: separate initial information gathering from resolution of triggering inputs in dataflow computation
When master/slave configurations are involved, initial_information actually does
nothing for slaves (they are processed with the master, added a comment to
explain why). But then, we call triggering_inputs, which may call done_port_info
on output ports for which there are no triggering inputs.

Then when the master gets processed, the slave's initial_task_information tries
to "create" the port info on a done port, which raises.

This worked so far because, by happenstance, master tasks were always added
before slave tasks in the final deployed network. The lazy deployments changed
the order and now this fails.